LayoutTests/http/wpt/opener/iframe-access-top-via-windowproxy.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

const wait_for_message = () => {
  return new Promise((resolve, reject) => {
    window.addEventListener('message', e => {
      resolve(e);
    }, { once: true });
  });
};

promise_test(async function(t) {
  let message = await wait_for_message();
  assert_equals(message.data, 'hello world!');
  let domains = await new Promise(resolve => {
    window.testRunner.getAndClearReportedWindowProxyAccessDomains(resolve);
  });
  assert_equals(domains.length, 0);
}, "Test that accessing WindowProxy within the same tab doesn't log.");
</script>

<iframe src="resources/send-message-to-top.html">
</iframe>
